add themes for JHEP and JCAP
JCAP (Journal of Cosmology and Astroparticle Physics) and JHEP (Journal
of High Energy Physics) are physics journals.  They use the LaTeX styles
`jcappub.sty` and `jheppub.sty`.

Add settings with figure widths based on the output of
```
linewidth=\uselengthunit{pt}\printlength{\linewidth}
\\
linewidth=\uselengthunit{in}\printlength{\linewidth}
```
within figures with in documents with these styles.

For JCAP these are
```
440.0pt
6.08948in
```

For JHEP these are
```
430.20639pt
5.95393in
```

https://jcap.sissa.it/jcap/help/JCAP_TeXclass.jsp
https://jhep.sissa.it/jhep/help/JHEP_TeXclass.jsp